Fix #1022 - when sending/recving data from sockets, check for count of actual bytes processed, and handle any remaining bytes if necessary.
Fix #1039 - docstring formatting.

Move ptvsd module from 'Visual Studio\Common7\IDE\...' to 'Program Files\Python Tools for Visual Studio'.

Pass process ID from debuggee to debugger when attaching.

Cancel pending request completion in REPL if connection drops.

Add the ability to use ptvsd as a command line tool to launch other scripts directly without the need to edit their code to import/enable_attach.

+++ Release/Product/Python/PythonTools/PythonTools/Repl/DebugReplThreadCommand.cs
@@ -37,8 +37,8 @@ namespace Microsoft.PythonTools.Repl {
                 if (string.IsNullOrEmpty(arguments)) {
                     eval.DisplayActiveThread();
                 } else {
-                    int id;
-                    if (int.TryParse(arguments, out id)) {
+                    long id;
+                    if (long.TryParse(arguments, out id)) {
                         eval.ChangeActiveThread(id, true);
                     } else {
                         window.WriteError(String.Format("Invalid arguments '{0}'. Expected thread id.", arguments));
